First attempt at adding Family objects to the project
I've attempted to add REST endpoints for creating Family objects and
dynamically adding and removing Persons to/from them.  Currently,
changes to the Family's list of Persons are not persisted between API
calls.  A Person will be successfully added, but then a subsequent call
to list all Persons in the family will result in an empty list.